% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign2Dgroups.R
\name{assign2Dgroups}
\alias{assign2Dgroups}
\title{Assign Group Membership for a Spectra2D Object}
\usage{
assign2Dgroups(spectra, gr.crit = NULL)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} missing \code{spectra$groups}.}

\item{gr.crit}{As per \code{\link{files2Spectra2DObject}}.}
}
\value{
A \emph{complete} object of S3 class \code{\link{Spectra2D}}.  This
function is the last internal step in creating a \code{Spectra2D} object.
Until this function has done its job, an object of class
\code{\link{Spectra2D}} will not pass checks as the assembly is not complete
(see \code{\link{chkSpectra2D}}).
}
\description{
A utility function which looks for \code{gr.crit} in the spectra/file names
and assigns group membership. Warnings are given if there are spectra/file names
that don't match entries in \code{gr.crit} or there are entries in
\code{gr.crit} that don't match any spectra/file names.  An internal function, not
generally called by the user.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\seealso{
\code{\link{files2Spectra2DObject}} for details;
\code{\link{sumGroups2D}} to see the outcome.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
