% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankPeaks2D.R
\name{blankPeaks2D}
\alias{blankPeaks2D}
\title{Blank Peaks in a Spectra2D Object}
\usage{
blankPeaks2D(spectra, remF2 = NULL, remF1 = NULL)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} from which to
remove selected frequencies.}

\item{remF2}{A formula giving the range of frequencies to be set to \code{NA}.  May include
"low" or "high" representing the extremes of the spectra.  See the examples.}

\item{remF1}{As for \code{remF2}.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
This function sets peaks at specified frequencies in a \code{\link{Spectra2D}}
object to \code{NA}. This effectively removes these peaks from calculations of
contours which can speed things up and clarifies the visual presentation of data.
This function is useful for removing regions with large
interfering peaks (e.g. the water peak in 1H NMR), or regions that are primarily
noise.
}
\details{
Note that the \code{\link[multiway]{parafac}} function does not allow \code{NA}
in the input data matrices.
}
\examples{

data(MUD1)
plotSpectra2D(MUD1, lvls = c(0.25, 0.5, 0.75),
  main = "MUD1 Spectrum 1: Complete Data Set")

# Note for this low resolution data set, setting
# remF2 = 3 ~ 4 results in no change due to rounding.

MUD1a <- blankPeaks2D(MUD1, remF2 = 2 ~ 4)
sumSpectra2D(MUD1a)
plotSpectra2D(MUD1a, , lvls = c(0.25, 0.5, 0.75),
  main = "Blank F2 2 ~ 4")

MUD1b <- blankPeaks2D(MUD1, remF2 = low ~ 5)
sumSpectra2D(MUD1b)
plotSpectra2D(MUD1b, , lvls = c(0.25, 0.5, 0.75),
  main = "Blank F2 low ~ 5")

MUD1c <- blankPeaks2D(MUD1, remF2 = 6 ~ high, remF1 = 4 ~ 13)
sumSpectra2D(MUD1c)
plotSpectra2D(MUD1c, lvls = c(0.25, 0.5, 0.75),
  main = "Blank F2 6 ~ high & F1 4 ~ 13")

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
