% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerSpectra2D.R
\name{centerSpectra2D}
\alias{centerSpectra2D}
\title{Center a Spectra2D Object Along the Samples Dimension}
\usage{
centerSpectra2D(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} from which to
extract peaks.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
This function centers a \code{Spectra2D} object along the samples dimension.
}
\examples{

data(MUD1)
tst <- centerSpectra2D(MUD1)

}
\references{
Need that article!
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
