% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpec2D-package.R
\docType{package}
\name{ChemoSpec2D-package}
\alias{ChemoSpec2D-package}
\alias{ChemoSpec2D}
\title{Exploratory Chemometrics for 2D Spectroscopy}
\description{
A collection of functions for working with 2D spectra,
developed with NMR in mind, but other types of 2D spectroscopy may also
be analyzed. ChemoSpec2D takes many of its cues from ChemoSpec
and tries to create consistent graphical output and be very user friendly.
Tools to plot 2D spectra and carry out PARAFAC are the main features.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\author{
Bryan A. Hanson.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{multivariate}
\keyword{package}
