% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkSpectra2D.R
\name{chkSpectra2D}
\alias{chkSpectra2D}
\title{Verify the Integrity of a Spectra2D Object}
\usage{
chkSpectra2D(spectra, confirm = FALSE)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} to be checked.}

\item{confirm}{Logical indicating whether or not to write the results to the
console, as would be desirable for interactive use.}
}
\value{
None; messages will be printed at the console if there is a problem.
}
\description{
Utility function, used by every function that has a \code{Spectra2D} object
as an argument, to verify that the structure of a \code{\link{Spectra2D}}
object (an S3 object) is internally consistent. This function can also be used
after manual editing of a \code{\link{Spectra2D}} object.  However,
in most cases rather than directly manipulating a \code{\link{Spectra2D}}
object, one should manipulate it via \code{\link{removeGroup2D}},
\code{\link{removeSample2D}}, \code{\link{normSpectra2D}},
or \code{\link{removeFreq2D}}.
}
\details{
This function is similar in spirit to \code{\link{validObject}} in the S4
world.  When used at the console, and the object is OK, no message is
written unless \code{confirm = TRUE}.  At the console, if there is a
problem, messages are issued regardless of the value of \code{confirm}.
When used in a function, this function is silent (assuming \code{confirm =
FALSE}) unless there is a problem.
}
\examples{

data(MUD1)
chkSpectra2D(MUD1, confirm = TRUE)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{classes}
\keyword{utilities}
