% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonExportedHelperFunctions.R
\name{.check4Gaps}
\alias{.check4Gaps}
\title{Check for Discontinuities (Gaps) in a Vector}
\usage{
.check4Gaps(x, tol = 0.01)
}
\arguments{
\item{x}{A numeric vector to be checked for gaps.}

\item{tol}{A number indicating the tolerance for checking to see if the step
between successive \code{x} values are the same.  Depending upon how the
\code{x} values are stored and rounded, you may need to change the value of
\code{tol} to avoid flagging trivial "gaps".}
}
\value{
A data frame giving the data chunks found, with one chunk per line.
\item{beg.freq }{The first frequency value in a
given data chunk.} \item{end.freq }{The last frequency value in a given data
chunk.} \item{size }{The length (in frequency units) of the data chunk.}
\item{beg.indx }{The index of the first frequency value in the data chunk.}
\item{eng.indx }{The index of the last frequency value in the data chunk.}
}
\description{
The basic procedure is to compare x[n + 1] - x[n] for successive values of
n.  When this value jumps, there is a gap which is flagged. \code{beg.indx}
and \code{end.indx} will always be contiguous as indices must be; it is the
\code{x} values that jump or have the gap.  The indices are provided as they
are more convenient in some programming contexts.  If not assigned, the
result appears at the console.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
