% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Esters.R
\docType{data}
\name{Esters}
\alias{Esters}
\title{Ester Data Sets}
\format{The data is stored as a \code{\link{Spectra2D}} object.}
\source{
Created using various tools.  Contact the author for a script if
interested.
}
\description{
This data is composed of simulated 300 MHz NMR COSY spectra.
The samples are methyl, ethyl, propyl, and butyl esters
of acetic, propanoic, butanoic, pentanoic and hexanoic acids.
Not all combinations of alcohols and acids are present.
There are 18 samples total.
}
\examples{
data(Esters)
sumSpectra2D(Esters)
sumGroups2D(Esters)

Esters2 <- removeGroup2D(Esters, "butyl") # remove the butyl esters
sumSpectra2D(Esters2)

Esters3 <- removeSample2D(Esters, "Pro") # remove the propanoates
sumSpectra2D(Esters3)

}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{datasets}
