% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessLvls.R
\name{guessLvls}
\alias{guessLvls}
\title{Automatically Select Contour Levels for 2D NMR}
\usage{
guessLvls(M, ...)
}
\arguments{
\item{M}{A matrix.}

\item{\dots}{Additional parameters to be passed to \code{\link{calcLvls}}.}
}
\value{
A vector of the suggested contour levels (invisibly).
}
\description{
Given a matrix for which one wants to make a contour plot, this
function will propose levels that are likely to work well
for 2D NMR data.  Uses \code{\link{calcLvls}} but makes
adjustments given the nature of 2D NMR data (i.e. this function
uses \code{mode = "log"} and removes the lowest contour level).
}
\examples{

set.seed(9)
MM <- matrix(rnorm(100, sd = 2), nrow = 10) # test data
guessLvls(MM, showHist = TRUE) # showHist passes through to calcLvls
guessLvls(abs(MM), showHist = TRUE)

}
\seealso{
Please see \code{\link{calcLvls}}.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utility}
