% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hats_alignSpectra2D.R
\name{hats_alignSpectra2D}
\alias{hats_alignSpectra2D}
\title{Align the Spectra in a Spectra2D Object using the HATS algorithm.}
\usage{
hats_alignSpectra2D(spectra, maxF2 = NULL, maxF1 = NULL,
  thres = 0.98, iter = 10, stopWhen = c(0.02, 5), plot = FALSE,
  debug = 0, method = "CA")
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{maxF2}{Integer.  The most extreme positive \code{F2step} to allow during the
alignment process (units are data points).  The search stops if this value is exceeded.}

\item{maxF1}{Integer.  As for \code{maxF2}, but for F2.}

\item{thres}{Numeric. The stopping criterion. Once the objective function reaches
this value the search stops.  The objective function is the cosine of the angle
between the unstacked spectra, so \code{thres} should be on [0 \ldots 1].}

\item{iter}{Integer. The maximum number of iterations.}

\item{stopWhen}{Length two numeric vector.  The first value is a change in the
objective function. Let's call it \code{v}.  The second value is a number of data points
(let's call it \code{n}, coerced to integer). The search process is stopped when the average
*change* of the most recent \code{n} values of the objective function falls below \code{v}.}

\item{plot}{Logical. Shall a plot of the alignment progress be made?  The plot is useful for
diagnostic purposes.  Every step of the alignment has a corresponding plot so you should
probably direct the output to a pdf file.}

\item{debug}{Integer.  Values >= 1 give messages about alignment progress.  Values >= 2
print the merge matrix from the \code{hclust} object, which serves as the guide tree.
For values >= 2 the same information is presented on the diagnostic plots.}

\item{method}{Character. Currently only \code{method = "CA"} is available which uses
the HATS algorithm plus a coordinate ascent method to align the spectra.
Use \code{plot = TRUE} to see this in action.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
Align the spectra in a \code{\link{Spectra2D}} object using an implementation
of the HATS algorithm described by Robinette \emph{et al.}.  Currently, only
global, not local, alignment is carried out.  The algorithm
used here assumes the shift needed to align the spectra is relatively small and
performs well in such cases. If some spectra are significantly shifted, the approach
here does not perform that well.  More data sets are needed for testing to improve the
algorithm details, please share!
}
\references{
Roughly follows the algorithm described in Robinette et al. 2011
           \emph{Anal. Chem.} vol. 83, 1649-1657 (2011) dx.doi.org/10.1021/ac102724x
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{multivariate}
