% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LofC.R
\name{LofC}
\alias{LofC}
\title{Create a List of Colors}
\usage{
LofC(cols, n = 1L, mode = 1L)
}
\arguments{
\item{cols}{Integer or Character.  If integer, \code{n} colors will be chosen from the standard color
scheme used by \code{ChemoSpec2D}.  Otherwise a character vector of color designations.}

\item{n}{Integer. The number of times to replicate the colors.}

\item{mode}{Integer. If \code{mode = 1L}, the values in \code{cols} will be replicated in each list entry.
If \code{mode = 2L}, the length of \code{cols} must equal \code{n} and the first list entry will 
be \code{n} values of the first color, the second list entry will be \code{n} values of the second color etc.}
}
\value{
A list of length \code{n}; each entry is a vector of colors.
}
\description{
When overlaying multiple 2D NMR spectra one needs to supply a vector of colors for each spectrum contour, as a list.
This is a convenience function which takes a vector of colors and copies it into a list, ready for
use  with \code{\link{plotSpectra2D}}.
}
\examples{
LofC(5, 1)
mycols <- c("red", "green", "blue")
LofC(mycols, 1)
LofC(mycols, 3, 2)
}
