% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUD.R
\docType{data}
\name{MUD1}
\alias{MUD1}
\title{Made Up 2D NMR-Like Data Sets}
\format{The data are stored as a \code{\link{Spectra2D}} object.}
\source{
Created from scratch. Contact the author for a script if interested.
}
\description{
Made Up Data that resemble a simple, HSQC-like 2D NMR data set.  Lean, low resolution 
and designed primarily to check graphics and test functions. \pkg{As this is made up
data, there is no underlying tri-linear structure and therefore one should NOT try to interpret
the output of \code{miaSpectra2D} or \code{pfacSpectra2D} run on this data}.
Real 2D NMR data sets are too large to be included in the package.
}
\details{
\code{MUD1} is composed of 10 samples in two groups.
}
\examples{

data(MUD1)
sumSpectra(MUD1)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{datasets}
