% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafacSpectra2D.R
\name{parafacSpectra2D}
\alias{parafacSpectra2D}
\title{PARAFAC Analysis of a Spectra2D Object}
\usage{
parafacSpectra2D(spectra, parallel = TRUE, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{parallel}{Logical.  Should parallel processing be used?}

\item{\dots}{Additional parameters to be passed to function parafac.}
}
\value{
An object of class \code{parafac}.
}
\description{
Carry out PARAFAC analysis of a \code{\link{Spectra2D}} object.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{multivariate}
