% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfacLoadings.R
\name{pfacLoadings}
\alias{pfacLoadings}
\title{Plot Loadings from a PARAFAC Analysis of a Spectra2D Object}
\usage{
pfacLoadings(spectra, pfac, which = 1, lvls = NULL, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{pfac}{An object of class \code{parafac}.}

\item{which}{An integer specifying the loading to plot.}

\item{lvls}{An integer specifying the levels at which to compute contours.
If \code{NULL}, values are computed using \code{chooseLvls}.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
Plots loadings from the PARAFAC analysis of a \code{\link{Spectra2D}} object.
The loadings are computed by multipling matrix \code{A} by matrix \code{B}
in the \code{parafac} object, for a given component.  This matrix has dimensions
F2 x F1.
}
\seealso{
Please see \code{\link{pfacSpectra2D}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
