% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfacScores.R
\name{pfacScores}
\alias{pfacScores}
\title{Plot Scores from a PARAFAC Analysis of a Spectra2D Object}
\usage{
pfacScores(spectra, pfac, which = c(1, 2), ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{pfac}{An An object of class \code{parafac}.}

\item{which}{An integer specifying the scores to plot.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
Plots a 2D spectrum stored in a \code{\link{Spectra2D}} object. UPDATE
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\seealso{
Please see \code{\link{pfacSpectra2D}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
