% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfacSpectra2D.R
\name{pfacSpectra2D}
\alias{pfacSpectra2D}
\title{PARAFAC Analysis of a Spectra2D Object}
\usage{
pfacSpectra2D(spectra, parallel = TRUE, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{parallel}{Logical.  Should parallel processing be used?
Unless you love waiting, you should use parallel processing.}

\item{\dots}{Additional parameters to be passed to function \code{\link[multiway]{parafac}}.}
}
\value{
An object of class \code{parafac}.
}
\description{
Carry out PARAFAC analysis of a \code{\link{Spectra2D}} object.
Function \code{\link[multiway]{parafac}} from \pkg{multiway} is used.
Because computation is slow, it is advisable to run this in batch
mode from the command line.
}
\references{
NEED to cite articles
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{multivariate}
