% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEngine.R
\name{plotEngine}
\alias{plotEngine}
\title{Plotting Engine for 2D Spectra}
\usage{
plotEngine(spectra, which = 1, lvls = NULL, colors = NULL, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{which}{An integer vector giving the spectra to be plotted.
Spectra are plotted in order so the last one is on top.}

\item{lvls}{A list of \code{length(which)}.  Each list element
should be a numeric vector giving the desired contour levels.
If any are \code{NULL}, values are computed using \code{chooseLvls}.}

\item{colors}{A vector of colors of \code{length(which)}.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
Side effect is a plot.
}
\description{
Plots one or more 2D spectra stored in a \code{\link{Spectra2D}} object.
This is the function that actually creates the plots requested
by several other functions.  Not intended to be called by the user.
Base graphics functions are used.  x and y axes dimensions are 0...1
par values should be adjusted before calling this function.
}
\examples{

data(MUD1)

plotSpectra2D(MUD1)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
