% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLoadings2D.R
\name{plotLoadings2D}
\alias{plotLoadings2D}
\title{Plot Loadings from a PARAFAC, MIA or POP Analysis of a Spectra2D Object}
\usage{
plotLoadings2D(
  spectra,
  so,
  load = 1,
  ref = NULL,
  load_lvls = NULL,
  ref_lvls = NULL,
  load_cols = NULL,
  ref_cols = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{so}{("Score Object") One of the following:
\itemize{
 \item An object of class \code{mia} produced by
  function \code{\link{miaSpectra2D}}.
 \item An object of class \code{pfac} produced by
  function \code{\link{pfacSpectra2D}}.
 \item An object of class \code{pop} produced by
  function \code{\link{popSpectra2D}}.
}}

\item{load}{An integer specifying the loading to plot.}

\item{ref}{An integer giving the spectrum in \code{spectra} to use
as a reference spectrum, which is plotted behind the loadings.
Defaults to \code{NULL} which does not plot a reference spectrum.}

\item{load_lvls}{A vector specifying the contour levels
for the loadings pseudo-spectrum.
If \code{NULL}, values are computed using \code{calcLvls}.}

\item{ref_lvls}{A vector specifying the levels at which to compute contours
for the reference spectrum.
If \code{NULL}, values are computed using \code{calcLvls}.}

\item{load_cols}{A vector specifying the colors for the contours in the laoding spectrum.
If \code{NULL}, defaults to a scheme of values
running from blue (low) to red (high), centered on green (zero).}

\item{ref_cols}{A vector specifying the colors for the contours in the reference
spectrum. If \code{NULL}, set to gray.}

\item{plot}{Logical.  Shall a plot be made?  Plotting large data sets can be slow.
Run the function with \code{plot = FALSE}, then use \code{\link{inspectLvls}}
to figure out desirable levels, then set \code{plot = TRUE}.}

\item{\dots}{Additional parameters to be passed to plotting functions.  For instance
\code{showGrid = TRUE}.}
}
\value{
The modified \code{Spectra2D} object is returned invisibly.
The loadings matrix will be appended with a sample of name of Loadings_x where
\code{x = load}. Side effect is a plot.
}
\description{
Computes (if necessary) and plots loadings from a PARAFAC, MIA or POP analysis of a
\code{\link{Spectra2D}} object. The loadings matrix has has dimensions
F1 x F2 and is a 2D pseudo-spectrum.  A reference spectrum may also be drawn.
}
\section{Scale}{

You can view the color scale for the plot via \code{\link{showScale}}.
}

\section{Levels & Colors}{

The number of levels and colors must match, and they are used 1 for 1.  If you
provide \code{n} colors, and no levels, the automatic calculation of levels may return
a number of levels other than \code{n}, in which case the function will override your colors and
assign new colors for the number of levels it computed (with a message).  To get
exactly what you want, specify both levels and colors in equal numbers.  Function
\code{\link{inspectLvls}} can help you choose appropriate levels.
If you specify more than one spectrum to plot, e.g. \code{which = c(1,2)}, then
arguments \code{lvls} and \code{cols} must be lists of levels and colors, one list
element for each spectrum to be plotted (if specified at all).  Two convenience functions exist to
make this process easier: \code{\link{LofL}} and \code{\link{LofC}}.
}

\seealso{
Please see \code{\link{pfacSpectra2D}}, \code{\link{miaSpectra2D}} or
\code{\link{popSpectra2D}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
