% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScree.R
\name{plotScree}
\alias{plotScree}
\title{Scree Plots of MIA Results for a Spectra2D Object}
\usage{
plotScree(results, ...)
}
\arguments{
\item{results}{A list as produced by \code{\link{miaSpectra2D}}.  Of particular interest are the
elements \code{C} containing the eigenvectors and \code{1c} containing the eigenvalues.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
Draw a scree plot to help choose the number of components in a MIA
(multivariate image analysis).
}
\references{
The idea for the alternative style scree plot produced here
came from the NIR-Quimiometria blog by jrcuesta, at
\url{https://nir-quimiometria.blogspot.com/2012/02/pca-for-nir-spectrapart-004-projections.html}
}
\seealso{
Please see \code{\link{miaSpectra2D}} for more information and examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
\keyword{multivariate}
