% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFreq2D.R
\name{removeFreq2D}
\alias{removeFreq2D}
\title{Remove Frequencies from a Spectra2D Object}
\usage{
removeFreq2D(spectra, remF2 = NULL, remF1 = NULL)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} from which to
extract peaks.}

\item{remF2}{A formula giving the range of frequencies to be extracted.  May include
"low" or "high" representing the extremes of the spectra.  See the examples.}

\item{remF1}{As for \code{remF2}.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
This function deletes specified frequencies from a \code{\link{Spectra2D}} object. This function
is useful for removing regions with large interfering peaks (e.g. the water peak in 1H NMR),
or regions that are primarily noise.
}
\examples{

data(MUD1)
plotSpectra2D(MUD1, lvls = c(0.25, 0.5, 0.75),
  main = "MUD1 Spectrum 1: Complete Data Set")

# Note for this low resolution data set, setting
# remF2 = 3 ~ 4 results in no change due to rounding.

MUD1a <- removeFreq2D(MUD1, remF2 = 2 ~ 4)
sumSpectra2D(MUD1a)
plotSpectra2D(MUD1a, , lvls = c(0.25, 0.5, 0.75),
  main = "Removed F2 2 ~ 4")

MUD1b <- removeFreq2D(MUD1, remF2 = low ~ 5)
sumSpectra2D(MUD1b)
plotSpectra2D(MUD1b, , lvls = c(0.25, 0.5, 0.75),
  main = "Removed F2 low ~ 5")

MUD1c <- removeFreq2D(MUD1, remF2 = 6 ~ high, remF1 = 4 ~ 13)
sumSpectra2D(MUD1c)
plotSpectra2D(MUD1c, lvls = c(0.25, 0.5, 0.75),
  main = "Removed F2 6 ~ high & F1 4 ~ 13")

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
