% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showScale.R
\name{showScale}
\alias{showScale}
\title{Display a pdf Version of the Contour Scale}
\usage{
showScale()
}
\description{
This function opens the files containing the contour scale in an appropriate viewer.
One file has a white background and the other a black background.
}
\examples{
\dontrun{
showScale()
}

}
