% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGroups2D.R
\name{sumGroups2D}
\alias{sumGroups2D}
\title{Summarize the Group Parameters of a Spectra2D Object}
\usage{
sumGroups2D(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} whose group
membership summary is desired.}
}
\value{
A data frame as follows.
        \item{group}{The name of the group.}
        \item{no.}{The number in the group.}
        \item{color}{The color assigned to the group.}
}
\description{
This function summarizes the group membership and descriptive parameters of
a \code{\link{Spectra2D}} object.
}
\examples{

data(MUD1)

sumGroups2D(MUD1)

}
\seealso{
To summarize the entire object, \code{\link{sumSpectra2D}}.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
