% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSpectra2D.R
\name{sumSpectra2D}
\alias{sumSpectra2D}
\title{Summarize a Spectra2D Object}
\usage{
sumSpectra2D(spectra, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{...}{Arguments to be passed downstream. If \code{sumSpectra2D} thinks there is a
gap between every data point, add the argument \code{tolF1 = xx} and/or 
\code{tolF2 = xx} which will pass through to
the underlying functions and override the internal guessing at appropriate values.}
}
\value{
None.  Results printed at console.
}
\description{
Provides a summary of a \code{\link{Spectra2D}} object, essentially a more
spectroscopist-friendly version of \code{str()}.
}
\details{
Prior to summarizing, \code{\link{chkSpectra2D}} is run with confirm = FALSE.
If there are problems, warnings are issued to the console and the summary is
not done.  The
\code{\link{Spectra2D}} object is also checked to see if it contains data elements
beyond what is required.  If so, these extra elements are reported to the
console.
}
\examples{

data(MUD1)
sumSpectra2D(MUD1)

MUD1a <- removeFreq2D(MUD1, remF2 = 2 ~ 4, remF1 = 10 ~ 15)
sumSpectra2D(MUD1a)

MUD1b <- removePeaks2D(MUD1, remF2 = 2 ~ 4, remF1 = 10 ~ 15)
sumSpectra2D(MUD1b)
plotSpectra2D(MUD1b, lvls = c(0.25, 0.5, 0.75),
  main = "Removed Peaks: F2 = 2 ~ 4, & F1 = 10 ~ 15")

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
