% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpec2D-package.R
\docType{package}
\name{ChemoSpec2D-package}
\alias{ChemoSpec2D-package}
\alias{ChemoSpec2D}
\title{Exploratory Chemometrics for 2D Spectroscopy}
\description{
A collection of functions for working with 2D spectroscopic data such as COSY and HSQC NMR 
spectra using appropriate chemometric techniques. ChemoSpec2D deploys methods aimed primarily
at classification of samples and the identification of spectral features which are important
in distinguishing samples from each other. Each 2D spectrum (a matrix) is treated as the unit of
observation, and thus the physical sample that went into the spectrometer corresponds to the
sample from a statistical perspective.  ChemoSpec2D takes many of its cues from ChemoSpec
and tries to create consistent graphical output and to be very user friendly.  Tools to
manipulate and plot 2D spectra and carry out PARAFAC are the main features.
}
\author{
Bryan A. Hanson.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}
