% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpec2D-package.R
\docType{package}
\name{ChemoSpec2D-package}
\alias{ChemoSpec2D-package}
\alias{ChemoSpec2D}
\title{Exploratory Chemometrics for 2D Spectroscopy}
\description{
Description: A collection of functions for exploratory chemometrics of 2D
spectroscopic data sets such as COSY and HSQC NMR spectra. ChemoSpec2D deploys
methods aimed primarily at classification of samples and the identification of
spectral features which are important in distinguishing samples from each other.
Each 2D spectrum (a matrix) is treated as the unit of observation, and thus the
physical sample in the spectrometer corresponds to the  sample from a statistical
perspective.  In addition to chemometric tools, a few tools are provided for
plotting 2D spectra, but these are not intended to replace the functionality
typically available on the spectrometer. ChemoSpec2D takes many of its cues from
ChemoSpec and tries to create consistent graphical output and to be very user friendly.
A vignette is available.
}
\author{
Bryan A. Hanson.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}
