% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkSpectra2D.R
\name{chkSpectra2D}
\alias{chkSpectra2D}
\title{Verify the Integrity of a Spectra2D Object}
\usage{
chkSpectra2D(spectra, confirm = FALSE)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} to be checked.}

\item{confirm}{Logical indicating whether or not to write the results to the
console, as would be desirable for interactive use.}
}
\value{
None; messages will be printed at the console if there is a problem.
}
\description{
Utility function to verify that the structure of a \code{\link{Spectra2D}}
object (an S3 object) is internally consistent. This function can be used
after manual editting of a \code{\link{Spectra2D}} object.  However,
in most cases rather than
directly manipulating a \code{\link{Spectra2D}} object, one should manipulate
it via \code{\link{removeGroup2D}}
or \code{\link{removeSample2D}}.
}
\details{
This function is similar in spirit to \code{\link{validObject}} in the S4
world.  When used at the console, and the object is OK, no message is
written unless \code{confirm = TRUE}.  At the console, if there is a
problem, messages are issued regardless of the value of \code{confirm}.
When used in a function, this function is silent (assuming \code{confirm =
FALSE}) unless there is a problem.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\seealso{
Please see \code{\link{ChemoSpec2D-package}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{classes}
\keyword{utilities}
