% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessLvls.R
\name{guessLvls}
\alias{guessLvls}
\title{Automatically Select Contour Levels 2D NMR}
\usage{
guessLvls(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
A vector of the suggested contour levels (invisibly).
}
\description{
Given a matrix for which one wants to make a contour plot, this
function will propose levels that are likely to work well
for 2D NMR data.  Uses \code{\link{calcLvls}} but makes
adjustments given the nature of 2D NMR data.
}
\examples{

set.seed(9)
MM <- matrix(runif(100, -1, 1), nrow = 10) # test data
print(guessLvls(MM))
print(guessLvls(abs(MM)))
print(guessLvls(-1*abs(MM)))

}
\seealso{
Please see \code{\link{calcLvls}}.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utility}
