% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2Dspectra.R
\name{import2Dspectra}
\alias{import2Dspectra}
\title{Import 2D Spectroscopic Data}
\usage{
import2Dspectra(file, fmt, npx, ...)
}
\arguments{
\item{file}{Character string giving the path to a file containing a 2D spectrum.}

\item{fmt}{Character string giving the format to use (currently limited).}

\item{npx}{Integer giving the number of data points in the F2 (x) dimension.}

\item{...}{Parameters to be passed to \code{\link{read.table}}.}
}
\value{
A numeric matrix.
}
\description{
This function imports a single matrix representing a 2D
spectroscopic data set.
}
\details{
If \code{fmt = "YXZ"} it is assumed that the columns contain Y, X and Z
values.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{import}
