% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import2Dspectra.R
\name{import2Dspectra}
\alias{import2Dspectra}
\title{Import 2D Spectroscopic Data}
\usage{
import2Dspectra(file, fmt, nF2, debug = 0, ...)
}
\arguments{
\item{file}{Character string giving the path to a file containing a 2D spectrum.}

\item{fmt}{Character string giving the format code to use.  Details below.}

\item{nF2}{Integer giving the number of data points in the F2 (x) dimension.
Note: If \emph{any} dimension is zero-filled you may need to study
the acquistion details to get the correct value for this argument.
This may be vendor-dependent.}

\item{debug}{Integer.  Applies to \code{fmt = "dx"} only.  See \code{\link[readJDX]{readJDX}}
for details.}

\item{...}{Parameters to be passed to \code{\link{read.table}}.}
}
\value{
A list with 3 elements:
\itemize{
  \item A matrix of the z values.  The no. of rows = \code{nF2} and the no. of columns
        follows from the size of the imported data.
  \item A vector giving the F2 (x) values.
  \item A vector giving the F1 (y) values.
 }
}
\description{
This function imports a single file (for instance, a csv) containing a 2D
spectroscopic data set.  The current version handles various types of
ASCII text files as well as a few other types.
This function is called by \code{\link{files2Spectra2DObject}}
and is exported and documented to assist in developing new format codes.
}
\section{NA}{

Here are some other format codes you can use:
\itemize{
  \item \code{"SimpleM"}.  Imports matrices composed of z values.  The F2 and F1 values
        are created from the dimension of the matrix.  After import, you will have to manually
        convert the F2 and F1 values to ppm.  You may also have to transpose the matrices, or
        perhaps invert the order of the rows or columns.  Read via \code{read.table}.
  \item \code{"Btotxt"}.  This format imports Bruker data written to a file using the Bruker
        "totxt" command.  Tested with TopSpin 4.0.7.  This format is read via \code{readLines}
        and thus the \ldots argument does not apply.
  \item \code{"dx"}.  This format imports files written in the JCAMP-DX format, via
        package \code{\link[readJDX]{readJDX}}.
}
}

\section{Other Format Codes}{

Here are some other format codes you can use:
\itemize{
  \item \code{"SimpleM"}.  Imports matrices composed of z values.  The F2 and F1 values
        are created from the dimension of the matrix.  After import, you will have to manually
        convert the F2 and F1 values to ppm.  You may also have to transpose the matrices, or
        perhaps invert the order of the rows or columns.  Read via \code{read.table}.
  \item \code{"Btotxt"}.  This format imports Bruker data written to a file using the Bruker
        "totxt" command.  Tested with TopSpin 4.0.7.  This format is read via \code{readLines}
        and thus the \ldots argument does not apply.
  \item \code{"dx"}.  This format imports files written in the JCAMP-DX format, via
        package \code{\link[readJDX]{readJDX}}.
}
}

\author{
Bryan A. Hanson, DePauw University.
}
\keyword{import}
