% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectLvls.R
\name{inspectLvls}
\alias{inspectLvls}
\title{Inspect Levels for Contour Plots of Spectra2D Objects}
\usage{
inspectLvls(spectra, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{\dots}{Arguments to be passed downstream.}
}
\value{
A numeric vector giving the levels (invisibly).
}
\description{
Given a \code{Spectra2D} object, this function will assist in selecting levels
for preparing contour and image type plots.  The entire range of the data is
used in the histogram so you can choose levels that are suitable for any
particular spectrum.  Any of the arguments to \code{\link{calcLvls}} can be used
to capture the computed levels, or you can choose by inspection.
}
\examples{

data(MUD1)
inspectLvls(MUD1, ylim = c(0, 200), main = "All MUD1 Data, mode = even")
inspectLvls(MUD1, ylim = c(0, 200), mode = "NMR",  main = "All MUD1 Data, mode = NMR")

}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\keyword{utilities}
