% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miaSpectra2D.R
\name{miaSpectra2D}
\alias{miaSpectra2D}
\title{Multivariate Image Analysis (Tucker 1) of a Spectra2D Object}
\usage{
miaSpectra2D(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}
}
\value{
A list per \code{\link[ThreeWay]{pcasup1}}.  Of particular interest are the
  elements \code{C} containing the eigenvectors and \code{1c} containing the eigenvalues.
}
\description{
Carry out multivariate image analysis of a \code{\link{Spectra2D}} object
(multivariate image analysis is the same as a Tucker1 analysis).
Function \code{\link[ThreeWay]{pcasup1}} from package \pkg{ThreeWay} is used.
}
\examples{

data(MUD1)
res <- miaSpectra2D(MUD1)
miaScores(MUD1, res, main = "MIA Scores")

}
\references{
A. Smilde, R. Bro and P. Geladi
"Multi-way Analysis: Applications in the Chemical Sciences" Wiley (2004).
See especially Example 4.5.

P. Geladi and H. Grahn "Multivariate Image Analysis" Wiley (1996).
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{multivariate}
