% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUD.R
\docType{data}
\name{MUD}
\alias{MUD}
\title{Made Up 2D NMR-Like Data Sets}
\format{The data are stored as a \code{\link{Spectra2D}} object.}
\source{
Created from scratch. Contact the author for a script if interested.
}
\description{
Made Up 2D NMR-Like Data Sets
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{datasets}
