% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUD.R
\docType{data}
\name{MUD1}
\alias{MUD1}
\title{Made Up 2D NMR-Like Data Sets}
\format{The data are stored as a \code{\link{Spectra2D}} object.}
\source{
Created from scratch. Contact the author for a script if interested.
}
\description{
Made Up Data that resemble 2D NMR data sets.  They are very small and designed primarily
to demonstrate and test functions.  Real 2D NMR data sets are too large to be included
in the package.
}
\details{
\code{MUD1} is composed of 10 samples in two groups.  The data resembles an HSQC spectrum.
}
\examples{

data(MUD1)

sumSpectra2D(MUD1)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{datasets}
