% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfacScores.R
\name{pfacScores}
\alias{pfacScores}
\title{Plot Scores from a PARAFAC Analysis of a Spectra2D Object}
\usage{
pfacScores(spectra, pfac, which = c(1, 2), ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{pfac}{An object of class \code{parafac}.}

\item{which}{An integer specifying two scores to plot.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
Plots scores from the PARAFAC analysis of a \code{\link{Spectra2D}} object.
The scores are the values returned in matrix \code{C} in the \code{parafac}
object.  This matrix has dimensions no. samples x no. of requested components.
}
\seealso{
Please see \code{\link{pfacSpectra2D}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
