% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra2D.R
\name{plotSpectra2D}
\alias{plotSpectra2D}
\title{Plot Spectra2D Object}
\usage{
plotSpectra2D(spectra, which = 1, lvls = NULL, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{which}{An integer specifying which spectrum to plot.}

\item{lvls}{An integer specifying the levels at which to compute contours.
If \code{NULL}, values are computed.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
Plots a 2D spectrum stored in a \code{\link{Spectra2D}} object.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\seealso{
Please see \code{\link{ChemoSpec2D-package}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
