% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra2D.R
\name{plotSpectra2D}
\alias{plotSpectra2D}
\title{Plot Spectra2D Object}
\usage{
plotSpectra2D(spectra, which = 1, lvls = NULL, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{which}{An integer specifying which spectrum to plot.}

\item{lvls}{An integer specifying the levels at which to compute contours.
If \code{NULL}, values are computed using \code{chooseLvls}.}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
Side effect is a plot.
}
\description{
Plots a 2D spectrum stored in a \code{\link{Spectra2D}} object.
This is primarily for inspection and for preparation of final plots.
If you need to do extensive exploration, you should probably go back
to the spectrometer.
}
\examples{

data(MUD1)

plotSpectra2D(MUD1)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
