% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra2D.R
\name{plotSpectra2D}
\alias{plotSpectra2D}
\title{Plot a Spectra2D Object}
\usage{
plotSpectra2D(spectra, which = 1, lvls = NULL, cols = NULL,
  showNA = TRUE, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{which}{An integer specifying which spectrum to plot.}

\item{lvls}{A numeric vector specifying the levels at which to compute contours.
If \code{NULL}, values are computed using \code{\link{calcLvls}}.}

\item{cols}{A vector of valid color designations.
If \code{NULL}, defaults to a scheme of nine values
running from blue (low) to red (high), centered on green (zero).}

\item{showNA}{Logical. Should the locations of peaks removed by \code{\link{removePeaks2D}}
be shown?  If present, these are show by a red line at each frequency.}

\item{\ldots}{Additional parameters to be passed to the plotting routines.}
}
\value{
Side effect is a plot.
}
\description{
Plots a 2D spectrum stored in a \code{\link{Spectra2D}} object.
This is primarily for inspection and for preparation of final plots.
If you need to do extensive exploration, you should probably go back
to the spectrometer.
}
\section{Warning}{

One cannot remove frequencies from the interior of a 2D NMR data set and expect to get a meaningful
contour plot, because doing so puts unrelated peaks adjacent in the data set.
This would lead to contours being drawn that don't exist in the original data set.
This function will check for missing frequencies and stop if any are found.
}

\examples{

data(MUD1)

plotSpectra2D(MUD1, main = "MUD1", lvls = c(0.25, 0.5, 0.75))

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{hplot}
