% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFreq2D.R
\name{removeFreq2D}
\alias{removeFreq2D}
\title{Remove Frequencies from a Spectra2D Object}
\usage{
removeFreq2D(spectra, remF2 = NULL, remF1 = NULL)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} from which to
remove selected frequencies.}

\item{remF2}{A formula giving the range of frequencies to be removed.
See the examples at \code{\link{ChemoSpec2D-package}}.}

\item{remF1}{As for \code{remF2}.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
This function sets the specified frequencies from a \code{\link{Spectra2D}}
object to \code{NA}.  For instance, one might want to remove
regions lacking any useful
information (to reduce the data size), or remove regions with large
interfering peaks (e.g. the water peak in 1H NMR).
}
\examples{

data(MUD1)

MUD1a <- removeFreq2D(MUD1, remF2 = 3 ~ 5)
MUD1b <- removeFreq2D(MUD1, remF2 = low ~ 5)

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
