% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFreq2D.R
\name{removeFreq2D}
\alias{removeFreq2D}
\title{Remove Frequencies from a Spectra2D Object}
\usage{
removeFreq2D(spectra, remF2 = NULL, remF1 = NULL)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} from which to
extract peaks.}

\item{remF2}{A formula giving the range of frequencies to be extracted.  May include
"low" or "high" representing the extremes of the spectra.  See the examples.}

\item{remF1}{As for \code{remF2}.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
This function deletes specified frequencies from a \code{\link{Spectra2D}} object.
This function is useful for removing regions with large interfering peaks
(e.g. the water peak in 1H NMR), or regions that are primarily noise.
It can also be used to isolate particular regions for further study.
}
\section{Warning}{

One cannot remove frequencies from the interior of a 2D NMR data set and expect
to get a meaningful contour plot, because doing so puts unrelated peaks adjacent
in the data set. This would lead to contours being drawn that don't exist in the
original data set. However, one can remove data from the interior and run a PARAFAC
analysis on the result, using the spectrum as an abstract object (that is, the
spectrum may not plottable, but the resulting scores are still meaningful).
}

\examples{

data(MUD1)
plotSpectra2D(MUD1, which = 7, lvls = seq(-1, 1, by = 0.2),
  main = "MUD1 Sample 7: Complete Data Set")

MUD1a <- removeFreq2D(MUD1, remF2 = 2 ~ 4)
sumSpectra2D(MUD1a) # cannot plot, results would be misleading

MUD1b <- removeFreq2D(MUD1, remF1 = low ~ 15)
sumSpectra2D(MUD1b)
plotSpectra2D(MUD1b, which = 7, lvls = seq(-1, 1, by = 0.2),
  main = "MUD1 Sample 7\\nRemoved Frequencies: F1 low ~ 15")

MUD1c <- removeFreq2D(MUD1, remF2 = low ~ 5)
sumSpectra2D(MUD1c)
plotSpectra2D(MUD1c, , which = 7, lvls = seq(-1, 1, by = 0.2),
  main = "MUD1 Sample 7\\nRemoved Frequencies: F2 low ~ 5")

MUD1d <- removeFreq2D(MUD1, remF2 = 6 ~ high, remF1 = 13 ~ 17)
sumSpectra2D(MUD1d) # not plotted, results would be misleading

}
\seealso{
\code{\link{removePeaks2D}} for another way to remove data.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
