% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftSpectra2D.R
\name{shiftSpectra2D}
\alias{shiftSpectra2D}
\title{Shift the Spectra in a Spectra2D Object}
\usage{
shiftSpectra2D(spectra, which = NULL, shiftF2 = 0L, shiftF1 = 0L,
  fill = "zero")
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{which}{An integer specifying which spectra to shift.  May be a vector.}

\item{shiftF2}{Integer.  The number of data points to shift along the F2 dimension.  See Details.}

\item{shiftF1}{As per \code{shiftF2}, but for the F1 dimension.}

\item{fill}{Character.  One of \code{"zero"} or \code{"rnorm"} describing how to fill
the empty rows or columns that are created during shifting.  If \code{rnorm}
you may wish to set the seed.}
}
\value{
An object of S3 class \code{\link{Spectra2D}}.
}
\description{
Shift the spectra in a \code{\link{Spectra2D}} object manually.  During shifting, some rows
or columns are thrown away and new rows or columns are introduced.  These new entries may
be filled with zeros, or random noise.
\itemize{
  \item (+) x-shift - shift right: trim right, fill left
  \item (-) x-shift - shift left: trim left, fill right
  \item (+) y-shift - shift up: trim top, fill bottom
  \item (-) y-shift - shift down: trim bottom, fill top
}
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{multivariate}
