% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGroups2D.R
\name{sumGroups2D}
\alias{sumGroups2D}
\title{Summarize the Group Parameters of a Spectra2D Object}
\usage{
sumGroups2D(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}} whose group
membership information is desired.}
}
\value{
A data frame as follows.  Note that if there are groups with no
members (due to previous use of \code{\link{removeSample2D}}), these are
dropped.  \item{group}{The name of the group.} \item{no.}{The number in the
group.}
}
\description{
This function summarizes the group membership and descriptive parameters of
a \code{\link{Spectra2D}} object.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec2D}
}
\seealso{
To summarize the entire object, \code{\link{sumSpectra2D}}.

Please see \code{\link{Esters}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
