% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSpectra2D.R
\name{sumSpectra2D}
\alias{sumSpectra2D}
\title{Summarize a Spectra2D Object}
\usage{
sumSpectra2D(spectra, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra2D}}.}

\item{...}{Arguments to be passed downstream.}
}
\value{
None.  Results printed at console.
}
\description{
Provides a summary of a \code{\link{Spectra2D}} object, essentially a more
spectroscopist-friendly version of \code{str()}.
}
\details{
Prior to summarizing, \code{\link{chkSpectra2D}} is run with confirm = FALSE.
If there are problems, warnings are issued to the console and the summary is
not done.  The
\code{\link{Spectra2D}} object is checked to see if it contains data elements
beyond what is required.  If so, these extra elements are reported to the
console.
}
\seealso{
Please see \code{\link{ChemoSpec2D-package}} for examples.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
